requirements.txt
wheel==0.44.0
apache-airflow==2.10.2
keras==3.6.0
tensorflow==2.18.0
azure-synapse-artifacts==0.19.0
azure-storage-blob==12.23.1
azure-identity==1.19.0
pandas==2.2.3
numpy==2.0.2
matplotlib==3.9.2
seaborn==0.13.2
sqlalchemy==1.4.54
python-dotenv==1.0.1
ipykernel==6.29.5
psycopg2-binary==2.9.10
streamlit==1.40.1
statsmodels==0.14.4
scikit-learn==1.5.2